+++ project/app/src/main/java/org/owntracks/android/support/OSSRequirementsChecker.kt
@@ -20,6 +20,15 @@ open class OSSRequirementsChecker @Inject constructor(open val context: Context)
           ContextCompat.checkSelfPermission(context, Manifest.permission.ACCESS_COARSE_LOCATION) ==
               PackageManager.PERMISSION_GRANTED
 
+  override fun hasBackgroundLocationPermission(): Boolean =
+      if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
+        ContextCompat.checkSelfPermission(
+            context, Manifest.permission.ACCESS_BACKGROUND_LOCATION) ==
+            PackageManager.PERMISSION_GRANTED
+      } else {
+        true
+      }
+
   override fun isLocationServiceEnabled(): Boolean =
       (context.getSystemService(Context.LOCATION_SERVICE) as LocationManager?)?.run {
         LocationManagerCompat.isLocationEnabled(this)