+++ project/app/src/main/java/org/owntracks/android/support/RequirementsChecker.kt
@@ -3,6 +3,8 @@ package org.owntracks.android.support
 interface RequirementsChecker {
   fun hasLocationPermissions(): Boolean
 
+  fun hasBackgroundLocationPermission(): Boolean
+
   fun isLocationServiceEnabled(): Boolean
 
   fun isPlayServicesCheckPassed(): Boolean