+++ project/app/src/main/java/org/owntracks/android/ui/welcome/fragments/LocationPermissionFragment.kt
@@ -12,6 +12,7 @@ import javax.inject.Inject
 import org.owntracks.android.databinding.UiWelcomeLocationPermissionBinding
 import org.owntracks.android.preferences.Preferences
 import org.owntracks.android.support.RequirementsChecker
+import org.owntracks.android.ui.mixins.BackgroundLocationPermissionRequester
 import org.owntracks.android.ui.mixins.LocationPermissionRequester
 import org.owntracks.android.ui.welcome.WelcomeViewModel
 
@@ -24,7 +25,16 @@ class LocationPermissionFragment @Inject constructor() : WelcomeFragment() {
   @Inject lateinit var requirementsChecker: RequirementsChecker
 
   private val locationPermissionRequester =
-      LocationPermissionRequester(this, ::permissionGranted, ::permissionDenied)
+      LocationPermissionRequester(
+          this,
+          { preferences.userDeclinedEnableLocationPermissions = false },
+          { preferences.userDeclinedEnableLocationPermissions = true })
+
+  private val backgroundLocationPermissionRequester =
+      BackgroundLocationPermissionRequester(
+          this,
+          { preferences.userDeclinedEnableBackgroundLocationPermissions = false },
+          { preferences.userDeclinedEnableBackgroundLocationPermissions = true })
 
   override fun shouldBeDisplayed(context: Context): Boolean = true
 
@@ -37,41 +47,48 @@ class LocationPermissionFragment @Inject constructor() : WelcomeFragment() {
     binding =
         UiWelcomeLocationPermissionBinding.inflate(inflater, container, false).apply {
           uiFragmentWelcomeLocationPermissionsRequest.setOnClickListener {
-            requestLocationPermissions()
+            locationPermissionRequester.requestLocationPermissions(0, requireContext()) {
+              shouldShowRequestPermissionRationale(it)
+            }
+          }
+          uiFragmentWelcomeLocationBackgroundPermissionsRequest.setOnClickListener {
+            backgroundLocationPermissionRequester.requestLocationPermissions(requireContext()) {
+              false
+            }
           }
         }
     return binding.root
-  }
-
-  private fun requestLocationPermissions() {
-    locationPermissionRequester.requestLocationPermissions(0, requireContext()) {
-      shouldShowRequestPermissionRationale(it)
-    }
   }
 
   override fun onResume() {
     super.onResume()
     viewModel.setWelcomeState(
         if (requirementsChecker.hasLocationPermissions()) {
-          binding.uiFragmentWelcomeLocationPermissionsRequest.visibility = View.INVISIBLE
-          binding.uiFragmentWelcomeLocationPermissionsMessage.visibility = View.VISIBLE
           WelcomeViewModel.ProgressState.PERMITTED
         } else if (preferences.userDeclinedEnableLocationPermissions) {
           WelcomeViewModel.ProgressState.PERMITTED
         } else {
           WelcomeViewModel.ProgressState.NOT_PERMITTED
         })
-  }
 
-  private fun permissionGranted(@Suppress("UNUSED_PARAMETER") code: Int) {
-    preferences.userDeclinedEnableLocationPermissions = false
-    binding.uiFragmentWelcomeLocationPermissionsRequest.visibility = View.INVISIBLE
-    binding.uiFragmentWelcomeLocationPermissionsMessage.visibility = View.VISIBLE
-    viewModel.setWelcomeState(WelcomeViewModel.ProgressState.PERMITTED)
+    if (requirementsChecker.hasLocationPermissions() &&
+        !requirementsChecker.hasBackgroundLocationPermission()) {
+      binding.uiFragmentWelcomeLocationBackgroundPermissionsRequest.visibility = View.VISIBLE
+      binding.uiFragmentWelcomeLocationPermissionsRequest.visibility = View.INVISIBLE
+      binding.uiFragmentWelcomeLocationPermissionsMessage.visibility = View.INVISIBLE
+    } else if (requirementsChecker.hasLocationPermissions() &&
+        requirementsChecker.hasBackgroundLocationPermission()) {
+      binding.uiFragmentWelcomeLocationBackgroundPermissionsRequest.visibility = View.INVISIBLE
+      binding.uiFragmentWelcomeLocationPermissionsRequest.visibility = View.INVISIBLE
+      binding.uiFragmentWelcomeLocationPermissionsMessage.visibility = View.VISIBLE
+    } else {
+      binding.uiFragmentWelcomeLocationBackgroundPermissionsRequest.visibility = View.INVISIBLE
+      binding.uiFragmentWelcomeLocationPermissionsRequest.visibility = View.VISIBLE
+      binding.uiFragmentWelcomeLocationPermissionsMessage.visibility = View.INVISIBLE
+    }
   }
 
   private fun permissionDenied(@Suppress("UNUSED_PARAMETER") code: Int) {
     preferences.userDeclinedEnableLocationPermissions = true
-    viewModel.setWelcomeState(WelcomeViewModel.ProgressState.PERMITTED)
   }
 }